Add dtype support to FDN configuration and modules
Introduces a 'dtype' field to BaseConfig and propagates it throughout FDN creation and initialization. This is in line with flamo 0.2. Updates all relevant classes and functions to accept and use the specified data type for tensor operations, improving flexibility and precision control. Also updates dependencies and version in pyproject.toml.